Move some extension logic
