+++ heinlein/dataset/__init__.py
@@ -1,3 +1,4 @@
-from .dataset import Dataset, dataset_extension, load_current_config, load_dataset
+from .dataset import Dataset, load_current_config, load_dataset
+from .extension import dataset_extension
 
 __all__ = ["Dataset", "load_dataset", "load_current_config", "dataset_extension"]
