+++ heinlein/manager/manager.py
@@ -81,6 +81,9 @@ def get_external_implementation(name: str) -> Optional[ModuleType]:
                 f" separately. You can install it with `pip install heinlein[{name}]`."
             )
             return None
+        print(module)
+        a = getattr(module, name)
+        print(a)
         return getattr(module, name)
